% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_default_risk_weights}
\alias{add_default_risk_weights}
\title{Set the default weight of each metric to 1.}
\usage{
add_default_risk_weights(data)
}
\arguments{
\item{data}{risk metric data}
}
\description{
Set the default weight of each metric to 1.
}
\keyword{internal}
