% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{assess_dep_for_sigmoid}
\alias{assess_dep_for_sigmoid}
\title{Assess dependencies for sigmoid point}
\usage{
assess_dep_for_sigmoid(riskdata_results)
}
\arguments{
\item{riskdata_results}{- path to riskdata_results}
}
\value{
nested list with dependencies, Import count mean, 
and all dependency count and mean
}
\description{
This function calculates average number of dependencies to 
determine the dependency sigmoid point
}
\keyword{internal}
