% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_description_file_elements}
\alias{assess_description_file_elements}
\title{assess_description_file_elements}
\usage{
assess_description_file_elements(pkg_name, pkg_source_path)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{pkg_source_path}{- pkg_source_path - source path for install local}
}
\value{
- list - list with scores for description file elements
}
\description{
assess_description_file_elements
}
\keyword{internal}
