% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_examples}
\alias{assess_examples}
\title{Assess Rd files for example or examples}
\usage{
assess_examples(pkg_name, pkg_source_path)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{pkg_source_path}{- source path for install local}
}
\value{
has_examples - variable with score
}
\description{
Assess Rd files for example or examples
}
\keyword{internal}
