% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_export_help}
\alias{assess_export_help}
\title{assess_export_help}
\usage{
assess_export_help(pkg_name, pkg_source_path)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{pkg_source_path}{- pkg_source_path - source path for install local}
}
\value{
- export_help - variable with score
}
\description{
assess_export_help
}
\keyword{internal}
