% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_exports}
\alias{assess_exports}
\title{Assess exported functions to namespace}
\usage{
assess_exports(data)
}
\arguments{
\item{data}{pkg source path}
}
\description{
Assess exported functions to namespace
}
\keyword{internal}
