% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_news}
\alias{assess_news}
\title{Assess Rd files for news}
\usage{
assess_news(pkg_name, pkg_source_path)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{pkg_source_path}{- source path for install local}
}
\value{
has_news - variable with score
}
\description{
Assess Rd files for news
}
\keyword{internal}
