% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_news_current}
\alias{assess_news_current}
\title{Assess Rd files for news}
\usage{
assess_news_current(pkg_name, pkg_ver, pkg_source_path)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{pkg_ver}{- package version}

\item{pkg_source_path}{- source path for install local}
}
\value{
news_current - variable with score
}
\description{
Assess Rd files for news
}
\keyword{internal}
