% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_size_codebase}
\alias{assess_size_codebase}
\title{assess codebase size}
\usage{
assess_size_codebase(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{- source path for install local}
}
\value{
- size_codebase - numeric value between \code{0} (for small codebase) and \code{1} (for large codebase)
}
\description{
Scores packages based on its codebase size, 
as determined by number of lines of code.
}
\keyword{internal}
