% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{assess_vignettes}
\alias{assess_vignettes}
\title{Assess vignettes}
\usage{
assess_vignettes(pkg_name, pkg_source_path)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{pkg_source_path}{- source path for install local}
}
\value{
- has_vignettes - variable with score
}
\description{
Assess vignettes
}
\keyword{internal}
