% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{calc_dependencies}
\alias{calc_dependencies}
\title{Calculate dependency score}
\usage{
calc_dependencies(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{package source path}
}
\description{
Calculate dependency score
}
\keyword{internal}
