% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{calc_reverse_dependencies}
\alias{calc_reverse_dependencies}
\title{Calculate reverse dependency score}
\usage{
calc_reverse_dependencies(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{package source path}
}
\description{
Calculate reverse dependency score
}
\keyword{internal}
