% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_risk_profile}
\alias{calc_risk_profile}
\title{Calculate package risk profile}
\usage{
calc_risk_profile(risk_data)
}
\arguments{
\item{risk_data}{overall risk score}
}
\value{
risk_profile
}
\description{
Calculate package risk profile
}
\examples{
\dontrun{
# Toy dataset
toy_data <- data.frame(score = c(0.1, 0.2, 0.3, 0.4, 0.8, 1.2))

calc_risk_profile(toy_data)
}

}
