% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{check_ggproto}
\alias{check_ggproto}
\title{function to check value of ggproto}
\usage{
check_ggproto(value)
}
\arguments{
\item{value}{- ggproto functions from the target package}
}
\description{
function to check all classes of a ggproto function
}
\keyword{internal}
