% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_risk_weights}
\alias{check_risk_weights}
\title{Check that the provided weights are numeric and non-negative.}
\usage{
check_risk_weights(weights)
}
\arguments{
\item{weights}{vector with weight values}
}
\description{
Check that the provided weights are numeric and non-negative.
}
\keyword{internal}
