% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{check_suggested_exp_funcs}
\alias{check_suggested_exp_funcs}
\title{Function to check suggested exported functions}
\usage{
check_suggested_exp_funcs(pkg_name, pkg_source_path, suggested_deps)
}
\arguments{
\item{pkg_name}{- name of the target package}

\item{pkg_source_path}{- source of the target package}

\item{suggested_deps}{- dependencies in Suggests}
}
\value{
- data frame with results of Suggests check
}
\description{
This function checks the exported functions of target package
against the exported functions of Suggested dependencies to see if any 
Suggested packages should be in Imports in the DESCRIPTION file
}
