% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_R_folder.R
\name{contains_r_folder}
\alias{contains_r_folder}
\title{Check for Vignette Folder and .Rmd Files in a .tar File}
\usage{
contains_r_folder(package_dir)
}
\arguments{
\item{package_dir}{A character string specifying the path to the package to be checked.}
}
\value{
A logical value: \code{TRUE} if the 'R' folder exists, \code{FALSE} otherwise.
}
\description{
This function checks if a given .tar file contains a 'vignettes' folder
and if there is at least one .Rmd file within that folder. If both 'vignettes'
and 'inst/doc' folders exist, the function will return \code{FALSE}.
}
\details{
The function checks if the specified folder exists. If  'R'
folder does not exist, the function returns \code{FALSE}. 
If the R folder exists, the function returns \code{TRUE}.
}
\keyword{internal}
