% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download.R
\name{convert_abbreviation_to_number}
\alias{convert_abbreviation_to_number}
\title{Convert Abbreviated Numbers to Integers}
\usage{
convert_abbreviation_to_number(value)
}
\arguments{
\item{value}{A character string representing a number (e.g., '1.2M', '500K', '100').}
}
\value{
An integer representation of the numeric value.
}
\description{
Converts a string containing a number with optional 'K' or 'M' abbreviation into an integer.
}
\keyword{internal}
