% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_results.R
\name{create_empty_results}
\alias{create_empty_results}
\title{Create empty results}
\usage{
create_empty_results(pkg_name, pkg_ver, pkg_source_path, metadata)
}
\arguments{
\item{pkg_name}{package name}

\item{pkg_ver}{package version}

\item{pkg_source_path}{package source path}

\item{metadata}{list with metadata}
}
\value{
- list with empty risk profile values
}
\description{
This creates an empty results list
for all risk metrics
}
\keyword{internal}
