% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_tm.R
\name{create_empty_tm}
\alias{create_empty_tm}
\title{create empty tm}
\usage{
create_empty_tm(pkg_name)
}
\arguments{
\item{pkg_name}{- name of package}
}
\value{
empty_tm - list with no tm
}
\description{
create empty tm
}
\keyword{internal}
