% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{create_file_coverage_df}
\alias{create_file_coverage_df}
\title{Generate file coverage df}
\usage{
create_file_coverage_df(file_names, file_coverage, errors, notes)
}
\arguments{
\item{file_names}{- file names}

\item{file_coverage}{- test coverage for files}

\item{errors}{- test coverage errors}

\item{notes}{- test coverage notes}
}
\description{
Generates file coverage df when errors list created.
}
\keyword{internal}
