% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{create_items_matched}
\alias{create_items_matched}
\title{Create items matched}
\usage{
create_items_matched(extracted_functions, suggested_exp_func)
}
\arguments{
\item{extracted_functions}{- exported functions from the target package}

\item{suggested_exp_func}{- exported functions from a suggested package}
}
\description{
create a df with functions that match exported functions from
a suggested package
}
\keyword{internal}
