% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_risk_profile.R
\name{create_risk_profile}
\alias{create_risk_profile}
\title{Create risk profile}
\usage{
create_risk_profile()
}
\value{
- list with risk profile values
}
\description{
This creates a risk profile 
for all risk metrics
}
\keyword{internal}
