% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights_profile.R
\name{create_weights_profile}
\alias{create_weights_profile}
\title{Create weights profile}
\usage{
create_weights_profile()
}
\value{
- numeric vector with weights profile
}
\description{
This creates a specific weights profile 
for all risk metrics
}
\examples{
\donttest{
create_weights_profile()
}
}
