% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{doc_riskmetric}
\alias{doc_riskmetric}
\title{Run all relevant documentation riskmetric checks}
\usage{
doc_riskmetric(pkg_name, pkg_ver, pkg_source_path)
}
\arguments{
\item{pkg_name}{name of the package}

\item{pkg_ver}{version of the package}

\item{pkg_source_path}{path to package source code (untarred)}
}
\value{
raw riskmetric outputs
}
\description{
Run all relevant documentation riskmetric checks
}
\keyword{internal}
