% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{extract_ggproto_methods}
\alias{extract_ggproto_methods}
\title{function to extract ggproto methods}
\usage{
extract_ggproto_methods(obj)
}
\arguments{
\item{obj}{- ggproto env from the target package}
}
\description{
function to get all the methods of a ggproto function
}
\keyword{internal}
