% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports.R
\name{extract_package_name}
\alias{extract_package_name}
\title{Extract package name from package source path}
\usage{
extract_package_name(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\value{
package_name
}
\description{
Extract package name from package source path
}
\keyword{internal}
