% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioconductor_data.R
\name{fetch_bioconductor_releases}
\alias{fetch_bioconductor_releases}
\title{Fetch Bioconductor Release Announcements}
\usage{
fetch_bioconductor_releases()
}
\value{
An XML document from bioconductor version page.
}
\description{
This function retrieves the Bioconductor release announcements page and returns
its HTML content for further processing.
}
\examples{
\dontrun{
html_content <- fetch_bioconductor_releases()
}
}
