% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{find_reverse_dependencies}
\alias{find_reverse_dependencies}
\title{find reverse dependencies}
\usage{
find_reverse_dependencies(path)
}
\arguments{
\item{path}{pkg_ref path}
}
\description{
find reverse dependencies
}
\keyword{internal}
