% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{function_is_ggproto}
\alias{function_is_ggproto}
\title{function to check value of ggproto}
\usage{
function_is_ggproto(x)
}
\arguments{
\item{x}{- ggproto functions from the target package}
}
\description{
function to check if first argument inherits from any of the classes specified
}
\keyword{internal}
