% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_coverage_section}
\alias{generate_coverage_section}
\title{Generate Coverage Section}
\usage{
generate_coverage_section(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the Coverage section for the HTML report.
}
\keyword{internal}
