% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_deps_section}
\alias{generate_deps_section}
\title{Generate Dependencies Section}
\usage{
generate_deps_section(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the dependencies section for the HTML report.
}
\keyword{internal}
