% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_doc_metrics_section}
\alias{generate_doc_metrics_section}
\title{Generate Doc Metrics Section}
\usage{
generate_doc_metrics_section(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the documentation section for the HTML report.
}
\keyword{internal}
