% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_html_report}
\alias{generate_html_report}
\title{Generate HTML Report for Package Assessment}
\usage{
generate_html_report(assessment_results, output_dir)
}
\arguments{
\item{assessment_results}{List containing the results from risk_assess_pkg function.}

\item{output_dir}{Character string indicating the directory where the report will be saved.}
}
\value{
Path to the generated HTML report.
}
\description{
Generates an HTML report for the package assessment results using rmarkdown.
}
\examples{
\dontrun{
assessment_results <- risk_assess_pkg()
generate_html_report(assessment_results, output_dir = "path/to/save/report")
}
}
