% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_rcmd_check_section}
\alias{generate_rcmd_check_section}
\title{Generate RCMD Check Metrics Section}
\usage{
generate_rcmd_check_section(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the RCMD Check Metrics HTML table for the HTML report.
}
\keyword{internal}
