% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_risk_details}
\alias{generate_risk_details}
\title{Generate Risk Details}
\usage{
generate_risk_details(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the Risk Details table for the HTML report.
}
\keyword{internal}
