% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_risk_summary}
\alias{generate_risk_summary}
\title{Generate Risk Summary}
\usage{
generate_risk_summary(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the Risk Summary table for the HTML report.
}
\keyword{internal}
