% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_report.R
\name{generate_trace_matrix_section}
\alias{generate_trace_matrix_section}
\title{Generate Trace Matrix Section}
\usage{
generate_trace_matrix_section(assessment_results)
}
\arguments{
\item{assessment_results}{- input data}
}
\description{
Generates the Trace Matrix section for the HTML report.
}
\keyword{internal}
