% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traceability_matrix.R, R/get_exports.R
\name{get_exports}
\alias{get_exports}
\title{list all package exports}
\usage{
get_exports(pkg_source_path)

get_exports(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\value{
data.frame, with one column `exported_function`, that can be passed
  to all downstream map_* helpers

data.frame, with one column `exported_function`, that can be passed
  to all downstream map_* helpers
}
\description{
adapted from mpn.scorecard

adapted from mpn.scorecard
}
\keyword{internal}
