% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_func_descriptions.R
\name{get_func_descriptions}
\alias{get_func_descriptions}
\title{Get function descriptions}
\usage{
get_func_descriptions(pkg_name)
}
\arguments{
\item{pkg_name}{- name of the package}
}
\description{
get descriptions of exported functions
}
\keyword{internal}
