% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{get_pkg_creator_info}
\alias{get_pkg_creator_info}
\title{check for package creator}
\usage{
get_pkg_creator_info(pkg_name, pkg_source_path)
}
\arguments{
\item{pkg_name}{- package name}

\item{pkg_source_path}{- path to package}
}
\value{
- logical - package creator exists - `FALSE`
}
\description{
check for package creator
}
\keyword{internal}
