% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_metrics.R
\name{get_pkg_license}
\alias{get_pkg_license}
\title{Assess License}
\usage{
get_pkg_license(pkg_name, pkg_source_path)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{pkg_source_path}{- source path for install local}
}
\value{
- Authors related variable
}
\description{
Assess License
}
\keyword{internal}
