% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_name}
\alias{get_pkg_name}
\title{get package name for display}
\usage{
get_pkg_name(input_string)
}
\arguments{
\item{input_string}{- string containing package name}
}
\value{
pkg_disp - package name for display
}
\description{
get package name for display
}
\examples{
\donttest{
pkg_source_path <- "/home/user/R/test.package.0001_0.1.0.tar.gz"
pkg_disp_1 <- get_pkg_name(pkg_source_path)
print(pkg_disp_1)

pkg <- "TxDb.Dmelanogaster.UCSC.dm3.ensGene_3.2.2.tar.gz"
pkg_disp_2 <- get_pkg_name(pkg)
print(pkg_disp_2)
}

}
