% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{get_regular_function_body}
\alias{get_regular_function_body}
\title{get regular function bodies}
\usage{
get_regular_function_body(pkg_name, func_name)
}
\arguments{
\item{pkg_name}{- name of the package}

\item{func_name}{- current function}
}
\description{
create a tibble with regular function bodies
}
\keyword{internal}
