% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_data.R
\name{get_repo_owner}
\alias{get_repo_owner}
\title{Extract GitHub repository owner from links}
\usage{
get_repo_owner(links, pkg_name)
}
\arguments{
\item{links}{A character vector of GitHub URLs.}

\item{pkg_name}{A string representing the name of the package.}
}
\value{
A string containing the GitHub repository owner, or NA if not found.
}
\description{
Given a vector of GitHub links and a package name, this function finds the link
corresponding to the package and extracts the GitHub repository owner.
}
\keyword{internal}
