% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{get_s3_method}
\alias{get_s3_method}
\title{function to get S3 method}
\usage{
get_s3_method(generic, class, package)
}
\arguments{
\item{generic}{- function generic}

\item{class}{- function class}

\item{package}{- package name}
}
\description{
function to get S3 method from generic
}
\keyword{internal}
