% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suggested_exp_funcs.R
\name{get_suggested_exp_funcs}
\alias{get_suggested_exp_funcs}
\title{Function to get suggested exported functions}
\usage{
get_suggested_exp_funcs(data)
}
\arguments{
\item{data}{- all packages listed in the DESCRIPTION file}
}
\value{
- data with package names and exported functions
}
\description{
This function gets exported functions for all packages
in the Suggests section of the target package's DESCRIPTION file
}
