% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traceability_matrix.R, R/get_exports.R
\name{get_toplevel_assignments}
\alias{get_toplevel_assignments}
\title{list all top-level objects defined in the package code}
\usage{
get_toplevel_assignments(pkg_source_path)

get_toplevel_assignments(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\value{
A data.frame with the columns `function` and `code_script` with a row for
  every top-level object defined in the package.

A data.frame with the columns `function` and `code_script` with a row for
  every top-level object defined in the package.
}
\description{
adapted from mpn.scorecard

adapted from mpn.scorecard
}
\details{
This is primarily for getting all _functions_, but it also returns top-level
declarations, regardless of type. This is intentional, because we also want
to capture any global variables or anything else that could be potentially
exported by the package.

This is primarily for getting all _functions_, but it also returns top-level
declarations, regardless of type. This is intentional, because we also want
to capture any global variables or anything else that could be potentially
exported by the package.
}
\keyword{internal}
