% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports.R
\name{map_functions_to_docs}
\alias{map_functions_to_docs}
\title{Map all Rd files to the functions they describe}
\usage{
map_functions_to_docs(exports_df, pkg_source_path, verbose)
}
\value{
Returns the data.frame passed to `exports_df`, with a `documentation`
  column appended. This column will contain the path to the `.Rd` files in
  `man/` that document the associated exported functions.
}
\description{
adapted from mpn.scorecard
}
\keyword{internal}
