% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traceability_matrix.R, R/get_exports.R
\name{map_functions_to_scripts}
\alias{map_functions_to_scripts}
\title{Get all exported functions and map them to R script where they are defined}
\usage{
map_functions_to_scripts(exports_df, pkg_source_path, verbose)

map_functions_to_scripts(exports_df, pkg_source_path, verbose)
}
\arguments{
\item{exports_df}{data.frame with a column, named `exported_function`,
containing the names of all exported functions. Can also have other columns
(which will be returned unmodified).}

\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}

\item{verbose}{check for extra information}
}
\value{
A data.frame with the columns `exported_function` and `code_script`.

A data.frame with the columns `exported_function` and `code_script`.
}
\description{
adapted from mpn.scorecard

adapted from mpn.scorecard
}
\keyword{internal}
