% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_description_file.R
\name{modify_description_file}
\alias{modify_description_file}
\title{Modify the DESCRIPTION File in a R Package Tarball}
\usage{
modify_description_file(tar_file)
}
\arguments{
\item{tar_file}{A string representing the path to the `.tar.gz` file that contains the R package.}
}
\value{
A string containing the path to the newly created modified `.tar.gz` file.
}
\description{
This function recreate a `.tar.gz` R package file after modifying its `DESCRIPTION` file
by appending Config/build/clean-inst-doc: false parameter.
}
\examples{
\dontrun{
  modified_tar <- modify_description_file("path/to/mypackage.tar.gz")
  print(modified_tar)
}

}
