% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_host_info.R
\name{parse_package_info}
\alias{parse_package_info}
\title{Parse Package Information from CRAN Archive}
\usage{
parse_package_info(name)
}
\arguments{
\item{name}{A character string specifying the name of the package to fetch information for.}
}
\value{
A character string containing the raw HTML content of the package archive page, or `NULL`
if the request fails or the package is not found.
}
\description{
This function retrieves the package archive information from the CRAN Archive.
}
\examples{
\dontrun{
# Fetch package archive information for "dplyr"
result <- parse_package_info("dplyr")

print(result)

}
}
