% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{preprocess_func_full_name}
\alias{preprocess_func_full_name}
\title{function to preprocess func_full_name}
\usage{
preprocess_func_full_name(func_full_name)
}
\arguments{
\item{func_full_name}{- full name of the function}
}
\description{
function to extract package, class, and generic
}
\keyword{internal}
