% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggested_exp_funcs.R
\name{process_items_matched}
\alias{process_items_matched}
\title{process items matched}
\usage{
process_items_matched(items_matched, suggested_exp_func)
}
\arguments{
\item{items_matched}{- exported functions match the suggested package function}

\item{suggested_exp_func}{- exported functions from a suggested package}
}
\description{
create a df with functions that match exported functions from
a suggested package
}
\details{
This function extracts all the functions from the function body 
and filters them to keep only function calls
It also extracts all the valid function names, matches with the source package
and writes a message
}
\keyword{internal}
