% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assess_pkg.R
\name{risk_assess_pkg}
\alias{risk_assess_pkg}
\title{Assess package - simplified}
\usage{
risk_assess_pkg(path = NULL)
}
\arguments{
\item{path}{(optional) path of locally stored package source code}
}
\value{
list containing results - list containing metrics, covr, tm - trace matrix, and R CMD check
}
\description{
simplified input to assess package for risk metrics
}
\examples{
\dontrun{
risk_assess_package <- risk_assess_pkg()

OR

risk_assess_package <- risk_assess_pkg(path/to/package.tar.gz)
}
}
