% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_assess_pkg_lock_files.R
\name{risk_assess_pkg_lock_files}
\alias{risk_assess_pkg_lock_files}
\title{Process lock files}
\usage{
risk_assess_pkg_lock_files(input_data)
}
\arguments{
\item{input_data}{- path to a lock file}
}
\value{
assessment_results - nested list containing risk metric data
}
\description{
This function processes `renv.lock` and `pak.lock` files 
to produce risk metric data
}
\examples{
\dontrun{
  input_data <- ("path/to/mypak.lock")
  pak_results <- risk_assess_pkg_lock_files(input_data) 
  print(pak_results)
 } 
}
