% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_metrics.R
\name{run_coverage}
\alias{run_coverage}
\title{Run covr and potentially save results to disk}
\usage{
run_coverage(pkg_source_path, timeout = Inf)
}
\arguments{
\item{pkg_source_path}{package installation directory}

\item{timeout}{Timeout to pass to [callr::r_safe()] when running covr.}
}
\value{
list with total coverage and function coverage
}
\description{
Run covr and potentially save results to disk
}
\keyword{internal}
