% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_metrics.R
\name{run_covr}
\alias{run_covr}
\title{Run covr in subprocess with timeout}
\usage{
run_covr(path, timeout)
}
\arguments{
\item{path}{- path to source file}

\item{timeout}{- length of timeout - set to Inf}
}
\description{
Run covr in subprocess with timeout
}
\keyword{internal}
