% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_pkg.R
\name{set_up_pkg}
\alias{set_up_pkg}
\title{Creates information on package installation}
\usage{
set_up_pkg(dp, check_type = "1")
}
\arguments{
\item{dp}{data path and name for the package.}

\item{check_type}{basic R CMD check type - "1" CRAN R CMD check_type - "2"}
}
\value{
list with local package install
}
\description{
Creates information on package installation
}
\examples{
\dontrun{
set_up_pkg(path/to/package, "mypackage")
}
}
