% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_rcmdcheck_args.R
\name{setup_rcmdcheck_args}
\alias{setup_rcmdcheck_args}
\title{set up rcmdcheck arguments}
\usage{
setup_rcmdcheck_args(check_type = "1", build_vignettes)
}
\arguments{
\item{check_type}{basic R CMD check type - "1" CRAN R CMD check_type - "2"}

\item{build_vignettes}{Logical (T/F). Whether or not to build vignettes}
}
\value{
- list with rcmdcheck arguments
}
\description{
This sets up rcmdcheck arguments
}
\details{
{Some packages need to have build vignettes as a 
build argument as their vignettes structure is inst/doc or inst/docs}
}
\keyword{internal}
