% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_risk_weights}
\alias{standardize_risk_weights}
\title{Check weights values and standardize them.}
\usage{
standardize_risk_weights(data, weights)
}
\arguments{
\item{data}{risk metric data}

\item{weights}{vector with weight values}
}
\description{
Check weights values and standardize them.
}
\keyword{internal}
