% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pscore_results.R
\name{update_pscore_results}
\alias{update_pscore_results}
\title{update pscore results}
\usage{
update_pscore_results(results, pscore)
}
\arguments{
\item{results}{list with results}

\item{pscore}{results from open source risk metrics}
}
\value{
- list with updated risk result values
}
\description{
This updates results list
for pscore risk metrics
}
\keyword{internal}
