% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_results_doc_scores.R
\name{update_results_doc_scores}
\alias{update_results_doc_scores}
\title{update results doc_metrics}
\usage{
update_results_doc_scores(results, doc_scores)
}
\arguments{
\item{results}{list with results}

\item{doc_scores}{results from documentation risk metrics}
}
\value{
- list with updated risk result values
}
\description{
This updates results list
for documentation risk metrics
}
